+++ pkg/dev_compiler/tool/input_sdk/private/isolate_helper.dart
@@ -385,8 +385,8 @@ class _IsolateContext implements IsolateContext {
   }
 
   void handlePing(SendPort responsePort, int pingType) {
-    if (pingType == Isolate.IMMEDIATE ||
-        (pingType == Isolate.BEFORE_NEXT_EVENT && !_isExecutingEvent)) {
+    if (pingType == Isolate.immediate ||
+        (pingType == Isolate.beforeNextEvent && !_isExecutingEvent)) {
       responsePort.send(null);
       return;
     }
@@ -394,7 +394,7 @@ class _IsolateContext implements IsolateContext {
       responsePort.send(null);
     }
 
-    assert(pingType == Isolate.BEFORE_NEXT_EVENT);
+    assert(pingType == Isolate.beforeNextEvent);
     if (_scheduledControlEvents == null) {
       _scheduledControlEvents = new Queue();
     }
@@ -403,12 +403,12 @@ class _IsolateContext implements IsolateContext {
 
   void handleKill(Capability authentification, int priority) {
     if (this.terminateCapability != authentification) return;
-    if (priority == Isolate.IMMEDIATE ||
-        (priority == Isolate.BEFORE_NEXT_EVENT && !_isExecutingEvent)) {
+    if (priority == Isolate.immediate ||
+        (priority == Isolate.beforeNextEvent && !_isExecutingEvent)) {
       kill();
       return;
     }
-    assert(priority == Isolate.BEFORE_NEXT_EVENT);
+    assert(priority == Isolate.beforeNextEvent);
     if (_scheduledControlEvents == null) {
       _scheduledControlEvents = new Queue();
     }